+++ crates/api/src/term.rs
@@ -6,8 +6,11 @@ mod invert;
 mod is_compatible_with;
 pub(crate) mod num_traits;
 mod partial_eq;
+mod partial_ord;
 mod reducible;
 mod ucum_unit;
+
+use std::borrow::Cow;
 
 use crate::{Atom, Prefix};
 
@@ -130,6 +133,79 @@ impl Term {
     pub fn factor_as_u32(&self) -> Factor {
         self.factor.unwrap_or(1)
     }
+
+    /// Depending on the `Term`, its string representation could be anywhere from a `&'static str`
+    /// to a combination of all of its fields as a `String`. For those former cases, we want to
+    /// allow borrowing the `&'static str` to save on allocations.
+    ///
+    #[must_use]
+    pub fn as_cow_str(&self) -> Cow<'_, str> {
+        use crate::UcumSymbol;
+
+        if self.is_unity() && self.annotation.is_none() {
+            return Cow::Borrowed("1");
+        };
+
+        match (
+            self.factor,
+            self.prefix,
+            self.atom,
+            self.exponent,
+            self.annotation.as_deref(),
+        ) {
+            (None, None, None, None, None) => Cow::Borrowed(""),
+            (None | Some(1), None, None, None | Some(1), Some(ann)) => {
+                Cow::Owned(format!("{{{ann}}}"))
+            }
+            (None, None, Some(atom), None | Some(1), None) => Cow::Borrowed(atom.primary_code()),
+            (None, None, Some(atom), None | Some(1), Some(ann)) => {
+                Cow::Owned(format!("{atom}{{{ann}}}"))
+            }
+            (None, None, Some(atom), Some(exp), None) => Cow::Owned(format!("{atom}{exp}")),
+            (None, None, Some(atom), Some(exp), Some(ann)) => {
+                Cow::Owned(format!("{atom}{exp}{{{ann}}}"))
+            }
+            (None, Some(prefix), Some(atom), None | Some(1), None) => {
+                Cow::Owned(format!("{prefix}{atom}"))
+            }
+            (None, Some(prefix), Some(atom), None | Some(1), Some(ann)) => {
+                Cow::Owned(format!("{prefix}{atom}{{{ann}}}"))
+            }
+            (None, Some(prefix), Some(atom), Some(exp), None) => {
+                Cow::Owned(format!("{prefix}{atom}{exp}"))
+            }
+            (None, Some(prefix), Some(atom), Some(exp), Some(ann)) => {
+                Cow::Owned(format!("{prefix}{atom}{exp}{{{ann}}}"))
+            }
+            (Some(factor), None, None, None, None) => Cow::Owned(factor.to_string()),
+            (Some(factor), None, None, None, Some(ann)) => Cow::Owned(format!("{factor}{{{ann}}}")),
+            (Some(factor), None, Some(atom), None | Some(1), None) => {
+                Cow::Owned(format!("{factor}{atom}"))
+            }
+            (Some(factor), None, Some(atom), None | Some(1), Some(ann)) => {
+                Cow::Owned(format!("{factor}{atom}{{{ann}}}"))
+            }
+            (Some(factor), None, Some(atom), Some(exp), None) => {
+                Cow::Owned(format!("{factor}{atom}{exp}"))
+            }
+            (Some(factor), None, Some(atom), Some(exp), Some(ann)) => {
+                Cow::Owned(format!("{factor}{atom}{exp}{{{ann}}}"))
+            }
+            (Some(factor), Some(prefix), Some(atom), None | Some(1), None) => {
+                Cow::Owned(format!("{factor}{prefix}{atom}"))
+            }
+            (Some(factor), Some(prefix), Some(atom), None | Some(1), Some(ann)) => {
+                Cow::Owned(format!("{factor}{prefix}{atom}{{{ann}}}"))
+            }
+            (Some(factor), Some(prefix), Some(atom), Some(exp), None) => {
+                Cow::Owned(format!("{factor}{prefix}{atom}{exp}"))
+            }
+            (Some(factor), Some(prefix), Some(atom), Some(exp), Some(ann)) => {
+                Cow::Owned(format!("{factor}{prefix}{atom}{exp}{{{ann}}}"))
+            }
+            _ => unreachable!("Invalid Term: {self:?}"),
+        }
+    }
 }
 
 #[cfg(test)]
@@ -139,5 +215,58 @@ mod tests {
     #[test]
     fn validate_new_unity() {
         assert_eq!(UNITY.to_string(), "1");
+    }
+
+    #[test]
+    fn as_str_test() {
+        assert_eq!(UNITY.as_cow_str(), "1");
+
+        // None, None, None, None, None
+        assert_eq!(Term::default().as_cow_str(), "");
+
+        // None | Some(1), None, None, None, Some(ann)
+        assert_eq!(term!(annotation: "hi".to_string()).as_cow_str(), "{hi}");
+        assert_eq!(
+            term!(factor: 1, annotation: "hi".to_string()).as_cow_str(),
+            "{hi}"
+        );
+
+        // None, None, Some(atom), None | Some(1), None
+        assert_eq!(term!(Meter).as_cow_str(), "m");
+        assert_eq!(term!(Meter, exponent: 1).as_cow_str(), "m");
+
+        // None, None, Some(atom), None | Some(1), Some(ann)
+        assert_eq!(
+            term!(Meter, annotation: "hi".to_string()).as_cow_str(),
+            "m{hi}"
+        );
+        assert_eq!(
+            term!(Meter, exponent: 1, annotation: "hi".to_string()).as_cow_str(),
+            "m{hi}"
+        );
+
+        assert_eq!(term!(Meter, exponent: 2).as_cow_str(), "m2");
+        assert_eq!(term!(Meter, exponent: -1).as_cow_str(), "m-1");
+
+        assert_eq!(
+            term!(Meter, exponent: 2, annotation: "hi".to_string()).as_cow_str(),
+            "m2{hi}"
+        );
+
+        assert_eq!(term!(Kilo, Meter).as_cow_str(), "km");
+
+        assert_eq!(
+            term!(Kilo, Meter, annotation: "hi".to_string()).as_cow_str(),
+            "km{hi}"
+        );
+
+        assert_eq!(
+            term!(Kilo, Meter, exponent: 1, annotation: "hi".to_string()).as_cow_str(),
+            "km{hi}"
+        );
+        assert_eq!(
+            term!(Kilo, Meter, exponent: 2, annotation: "hi".to_string()).as_cow_str(),
+            "km2{hi}"
+        );
     }
 }